small change for flashcard route
